+++ src/shared/message/server_message/message_buffer.rs
@@ -31,13 +31,18 @@ impl MessageBuffer {
         self.ticks.last_mut()
     }
 
-    pub(super) fn insert(&mut self, mode: SendMode, channel_id: usize, message: SerializedMessage) {
+    pub(super) fn insert(
+        &mut self,
+        targets: SendTargets,
+        channel_id: usize,
+        message: SerializedMessage,
+    ) {
         let buffer = self
             .active_tick()
             .expect("`start_tick` should be called before buffering");
 
         buffer.messages.push(BufferedMessage {
-            mode,
+            targets,
             channel_id,
             message,
         });
@@ -57,8 +62,8 @@ impl MessageBuffer {
     ) -> Result<()> {
         for mut tick in self.ticks.drain(..) {
             for mut message in tick.messages.drain(..) {
-                match message.mode {
-                    SendMode::Broadcast => {
+                match message.targets {
+                    SendTargets::All => {
                         for (client, ticks) in
                             clients.iter().filter(|(e, _)| !tick.excluded.contains(e))
                         {
@@ -72,7 +77,7 @@ impl MessageBuffer {
                             }
                         }
                     }
-                    SendMode::BroadcastExcept(ignored_id) => {
+                    SendTargets::AllExcept(ignored_id) => {
                         for (client, ticks) in
                             clients.iter().filter(|(c, _)| !tick.excluded.contains(c))
                         {
@@ -90,7 +95,7 @@ impl MessageBuffer {
                             }
                         }
                     }
-                    SendMode::Direct(client_id) => {
+                    SendTargets::Single(client_id) => {
                         if let ClientId::Client(client) = client_id
                             && let Ok((_, ticks)) = clients.get(client)
                             && !tick.excluded.contains(&client)
@@ -136,7 +141,7 @@ impl TickMessages {
 }
 
 struct BufferedMessage {
-    mode: SendMode,
+    targets: SendTargets,
     channel_id: usize,
     message: SerializedMessage,
 }